main/requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiohttp==3.8.6
aiosignal==1.3.2
ajsonrpc==1.2.0
annotated-types==0.7.0
anyio==3.7.1
appnope==0.1.4
argcomplete==3.5.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.2
bitarray==2.9.3
bitstring==4.2.3
bleach==6.1.0
blinker==1.8.2
bottle==0.13.2
bs4==0.0.2
cattrs==24.1.2
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chex==0.1.89
click==8.1.7
colorama==0.4.6
comm==0.2.2
concurrent-iterator==0.2.6
contourpy==1.2.1
coremltools==8.1
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
ecdsa==0.19.0
esptool==4.8.1
etils==1.12.0
executing==2.1.0
ExifRead==3.0.0
fake-useragent==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.3
flatbuffers==25.2.10
flax==0.10.3
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.13.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.26.3
humanize==4.12.1
idna==3.10
importlib_resources==6.5.2
intelhex==2.3.0
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.5.1
jaxlib==0.5.1
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.8.0
Keras-Applications==1.0.8
kiwisolver==1.4.5
libclang==18.1.1
llvmlite==0.43.0
lxml==4.9.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.8.1
nose==1.3.7
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
openai==1.3.0
openai-whisper @ git+https://github.com/openai/whisper.git@173ff7dd1d9fb1c4fddea0d41d704cfefeb8908c
opencv-contrib-python-headless==4.11.0.86
opencv-python==4.8.1.78
opt_einsum==3.4.0
optax==0.2.4
optree==0.14.0
orbax-checkpoint==0.11.6
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.3.6
platformio==6.1.16
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.3.0
protobuf==4.25.6
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==24.9.0
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pyelftools==0.31
Pygments==2.18.0
pyparsing==3.1.2
pyserial==3.5
PySide6==6.7.3
PySide6_Addons==6.7.3
PySide6_Essentials==6.7.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
pytz==2024.1
PyWavelets==1.5.0
PyYAML==6.0.2
pyzmq==26.2.0
rawpy==0.19.1
reedsolo==1.7.0
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.25.0
semantic-version==2.10.0
Send2Trash==1.8.3
serpapi==0.1.5
sewar==0.4.6
shiboken6==6.7.3
simplejson==3.20.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
starlette==0.39.2
streamlink==7.0.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorstore==0.1.72
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.21.0
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.49.0
treescope==0.1.9
trio==0.26.2
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.30.6
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.17.2
wsproto==1.2.0
xgboost==1.6.2
yarl==1.18.3
zipp==3.21.0
